added new changes in public and admin api endpoints for products & categories managment

+++ client/src/services/productService.js
@@ -1,14 +1,14 @@
 import api from "./api";
 
 export const addProduct = (data) => {
-  return api.post("/product", data, {
+  return api.post("/admin/product", data, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
   });
 };
 export const updateProduct = (id, data) => {
-  return api.post("/product/" + id, data, {
+  return api.post("/admin/product/" + id, data, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -16,11 +16,11 @@ export const updateProduct = (id, data) => {
 };
 export const getProducts = (filters) => {
    const queryString = new URLSearchParams(filters).toString(); 
-  return api.get(`/product?${queryString}`);
+  return api.get(`/admin/product?${queryString}`);
 };
 export const deleteProduct = (id) => {
-  return api.delete(`/product/${id}`);
+  return api.delete(`/admin/product/${id}`);
 };
 export const restoreProduct = (id) => {
-  return api.put(`/product/${id}/restore`);
+  return api.put(`/admin/product/${id}/restore`);
 };
